First refactor to port-adapter design.
