+++ src/main/java/com/scmspain/application/services/TweetService.java
@@ -1,6 +1,6 @@
 package com.scmspain.application.services;
 
-import com.scmspain.infrastructure.configuration.entities.Tweet;
+import com.scmspain.infrastructure.entities.Tweet;
 import org.springframework.boot.actuate.metrics.writer.Delta;
 import org.springframework.boot.actuate.metrics.writer.MetricWriter;
 import org.springframework.stereotype.Service;
